Debugged edit/cancel-edit/add/hide/unhide functionalities

+++ src/main/java/app/ItemJ.java
@@ -6,10 +6,10 @@ public class ItemJ implements Serializable{
     private int id;
     private String name;
     private String description;
-    private int price;
+    private double price;
     private int quantity;
 
-    public ItemJ(int id, String name, String description, int price, int quantity) {
+    public ItemJ(int id, String name, String description, double price, int quantity) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -44,11 +44,11 @@ public class ItemJ implements Serializable{
         this.description = description;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
